move all tests to a single package
