+++ src/gametest/resources/fabric.mod.json
@@ -7,17 +7,17 @@
     "license": "LGPL-3.0",
     "entrypoints": {
         "fabric-gametest": [
-            "com.github.thedeathlycow.frostiful.test.icicle.IcicleBreakTests",
-            "com.github.thedeathlycow.frostiful.test.icicle.FrostTippedArrowTests",
-            "com.github.thedeathlycow.frostiful.test.sun_lichen.SunLichenCollisionTests",
-            "com.github.thedeathlycow.frostiful.test.sun_lichen.SunLichenPathfindingTests",
-            "com.github.thedeathlycow.frostiful.test.effects.RootedTests",
-            "com.github.thedeathlycow.frostiful.test.effects.BlockLightTemperatureTests",
-            "com.github.thedeathlycow.frostiful.test.frostologer.FrostologerDestroyHeatSourcesTests",
-            "com.github.thedeathlycow.frostiful.test.frostologer.ChillagerDropTests",
-            "com.github.thedeathlycow.frostiful.test.powder_snow_walkable.EntityPowderSnowTests",
-            "com.github.thedeathlycow.frostiful.test.damage.HotFloorTests",
-            "com.github.thedeathlycow.frostiful.test.loot.FrozenByFrostWandCriterionTest"
+            "com.github.thedeathlycow.frostiful.test.tests.IcicleBreakTests",
+            "com.github.thedeathlycow.frostiful.test.tests.FrostTippedArrowTests",
+            "com.github.thedeathlycow.frostiful.test.tests.SunLichenCollisionTests",
+            "com.github.thedeathlycow.frostiful.test.tests.SunLichenPathfindingTests",
+            "com.github.thedeathlycow.frostiful.test.tests.RootedTests",
+            "com.github.thedeathlycow.frostiful.test.tests.BlockLightTemperatureTests",
+            "com.github.thedeathlycow.frostiful.test.tests.FrostologerDestroyHeatSourcesTests",
+            "com.github.thedeathlycow.frostiful.test.tests.ChillagerDropTests",
+            "com.github.thedeathlycow.frostiful.test.tests.EntityPowderSnowTests",
+            "com.github.thedeathlycow.frostiful.test.tests.HotFloorTests",
+            "com.github.thedeathlycow.frostiful.test.tests.FrozenByFrostWandCriterionTest"
         ]
     }
 }
